Footer Bug
Footer “floats” above bottom of the page. When the cursor enters the
area beneath the footer, the background image appears. When the list
hover background image is activated, this background image “shines
through” beneath the footer element.

Also, when top padding is applied to the top of the footer element
styled with a background color, the footer occupying text leaves the
footer element and positions toward the bottom of the page.

A solution so far is to modify the method for image hover.